backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-groq
mysql-connector-python
python-dotenv
